+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Previously known as My Account Page Editor
 
 ## 1.1.7 / 2024-08-02
 - Removed page for Locations, Banners and Subsidiaries.
+- Fixed loading of ACC pages.
 
 ## 1.1.6 / 2024-08-01
 - New block: Touchpoints MicroSpec. Can also show single event data on a page.